Move detection logic from moveObjects to lookAhead
moveObjects is unable to detect moving boxes as they are not at the same
position in the final state.
Now moveObjects simply finds a free spot of the type found in lookAhead.

It's possible that this is instead a sign that we need to handle moving
boxes differently, e.g. just ignoring them and hoping to fix them by
waiting (inserting NoOp's) until they have passed.